Use title, abstract, and keywords for classification in consistent order
The set_paper() method of CSOClassifierSemantic takes either a string or
a dictionary. If passed a dictionary, it constructs a string. Previously
it iterated over all the dictionary's keys and concatenated the
corresponding values. This revision processes only the keys "title",
"abstract", and "keywords," consistent with the docstring, to avoid
unexpected behavior when other keys are present. Also, iterating over
dictionary keys happens either in arbitrary order (Python < 3.6) or in
order of insertion (>= 3.6). This revision concatenates the input text
in consistent order.