Only call set_paper() during classifier init for non-missing papers
Input text can be passed during instantiation of the class via the
"paper" argument of the init method, or later via the set_paper method.
If the former, the init method calls the set_paper method on the input
for preprocessing. Only do this if a paper is passed to the init method.